Implemented Distributed Strategy for Healthy Agents
Implemented a heuristic distributed strategy that gathers agents in
uniform circles. This strategy assumes full communication
between agents and shared coordinate system.
We also lift the restrictions of anonnymity and oblivious robots.
Still, agents can conceal one another which makes the problem though.

Remaining Work:
Add code that takes into account that agents can be conquered throughout
the game. We are only interested in communicating with healthy agents
and we need to remember that agents in our circle  can be conquered
and we need to react accordingly.